Add extension function to convert a utc time to an arbitrary local time
Normally, the DateTime.ToLocalTime() will work off of system time. This
new extension method takes an offset in seconds that will convert it to an
arbitrary local time.

This will be useful since the OpenWeatherMap api returns a UTC offset
in the response, so we can report local times for the cities that are
searched. :^)