Silence a compiler warning about CurrentCondition being null.
Specifically the compiler was complaining that CurrentCondition was not
set after the constructor completed. However, the important thing is
it was never null when it rendered the page.

Adding a question mark silences the compiler on this warning, however
it introduced a different warning that was fixed in the previous
commit.